Refactor StockService to add throws

+++ src/main/java/com/mandacarubroker/service/StockService.java
@@ -3,8 +3,12 @@ package com.mandacarubroker.service;
 import com.mandacarubroker.domain.stock.RequestStockDTO;
 import com.mandacarubroker.domain.stock.Stock;
 import com.mandacarubroker.domain.stock.StockRepository;
+import com.mandacarubroker.service.exceptions.CampoNuloException;
+import com.mandacarubroker.service.exceptions.CampoVazioException;
 import com.mandacarubroker.service.exceptions.ObjectNotFoundException;
+import com.mandacarubroker.service.exceptions.RegistroNaoEncontradoException;
 import jakarta.validation.*;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -31,25 +35,44 @@ public class StockService {
     }
 
     public Stock createStock(RequestStockDTO data) {
-        Stock novaAcao = new Stock(data);
-        validateRequestStockDTO(data);
-        return stockRepository.save(novaAcao);
+        try {
+            Stock novaAcao = new Stock(data);
+            validateRequestStockDTO(data);
+            return stockRepository.save(novaAcao);
+        }catch (ConstraintViolationException ex){
+            throw new CampoVazioException("Os campos symbol e companyName não podem estar vazios ou nulos.");
+        }
     }
 
     public Optional<Stock> updateStock(String id, Stock updatedStock) {
-        return stockRepository.findById(id)
-                .map(stock -> {
-                    stock.setSymbol(updatedStock.getSymbol());
-                    stock.setCompanyName(updatedStock.getCompanyName());
-                    double newPrice = stock.changePrice(updatedStock.getPrice(), true);
-                    stock.setPrice(newPrice);
+        Optional<Stock> existingStock = stockRepository.findById(id);
 
-                    return stockRepository.save(stock);
-                });
+        if (existingStock.isEmpty()) {
+            throw new ObjectNotFoundException("Registro não encontrado para o ID: " + id);
+        } else if (updatedStock.getSymbol() == null || updatedStock.getCompanyName() == null) {
+            throw new CampoNuloException("Os campos symbol e companyName não podem ser nulos.");
+        } else if (updatedStock.getSymbol().isBlank() || updatedStock.getCompanyName().isBlank()) {
+            throw new CampoVazioException("Os campos symbol e companyName não podem estar vazios.");
+        }
+        else{
+            return existingStock
+                    .map(stock -> {
+                        stock.setSymbol(updatedStock.getSymbol());
+                        stock.setCompanyName(updatedStock.getCompanyName());
+                        double newPrice = stock.changePrice(updatedStock.getPrice(), true);
+                        stock.setPrice(newPrice);
+
+                        return stockRepository.save(stock);
+                    });
+        }
     }
 
     public void deleteStock(String id) {
-        stockRepository.deleteById(id);
+        if (stockRepository.findById(id).isEmpty()) {
+            throw new ObjectNotFoundException("Registro não encontrado para o ID: " + id);
+        }else {
+            stockRepository.deleteById(id);
+        }
     }
 
     public static void validateRequestStockDTO(RequestStockDTO data) {
